requirements.txt
pillow>=5.0.0